added CaseUpdateCommand superclass and refactored add_hq_case_id_to_case (no logic change)
